requirements.txt
django
djangorestframework
djangorestframework_simplejwt
psycopg2-binary
dj-config-url
python-decouple
drf-yasg
django-cors-headers
pillow
django-filter
celery[redis]
loguru
coverage
model_bakery
geocoder
selenium
requests
webdriver_manager
fake-useragent
bs4
lxml
social-auth-app-django
channels
channels-redis
django-allauth